third_party/blink/web_tests/external/wpt/editing/run/multitest_9001-last-expected.txt

This is a testharness.js-based test.
Found 46 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=blue>[bar</font></a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=blue><a href=http://www.google.com>[bar</a></font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=brown>[bar</font></a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\"><font color=\\"#a52a2a\\">a</font></a>" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font color=brown><a href=http://www.google.com>[bar</a></font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><font color=black>[bar</font></a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\"><font color=\\"#000000\\">a</font></a>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<a href=http://www.google.com><u>[bar</u></a>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<u><a href=http://www.google.com>[bar</a></u>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<a href=\\"http://www.google.com\\">a</a>" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub><font size=2>[bar</font></sub>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>" but got "foo<span style=\\"font-size:small\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=2><sub>[bar</sub></font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"1\\"><sub>a</sub></font>" but got "foo<span style=\\"font-size:10.8333px\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "foo<sub><font size=3>[bar</font></sub>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"3\\"><sub>a</sub></font>" but got "fooa"
[FAIL] [["delete",""],["inserttext","a"]] "foo<font size=3><sub>[bar</sub></font>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>" but got "foo<span style=\\"font-size:13.3333px\\">a</span>"
[FAIL] [["delete",""],["inserttext","a"]] "<blockquote><font color=blue>[foo]</font></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><font color=\\"blue\\">a</font></blockquote>" but got "<blockquote><span style=\\"color:rgb(0, 0, 255)\\">a</span></blockquote>"
[FAIL] [["styleWithCSS","false"],["delete",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["insertparagraph",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["insertparagraph",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["insertparagraph",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["delete",""],["insertparagraph",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["insertparagraph",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["insertparagraph",""],["inserttext","a"]] "<div><b>[abc]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["insertparagraph",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["insertparagraph",""],["inserttext","d"]] "<div>abc<b>[def]</b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b>abc[]<br></b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b>abc[]<br></b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><i><b>abc[]<br></b></i></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><i><b>abc[]<br></b></i></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><i><b>abc[]</b><br></i></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><i><b>abc[]</b><br></i></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i></b></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><i style=\\"\\"><b>abc</b>d<br></i></div>" not in array ["<div><b><i>abc</i></b><i>d</i></div>", "<div><b><i>abc</i></b><i>d<br></i></div>", "<div><b><i>abc</i></b><i>d</i><br></div>"]
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i></b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i></b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]</i><br></b></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><i style=\\"\\"><b>abc</b>d</i><br></div>" not in array ["<div><b><i>abc</i></b><i>d</i></div>", "<div><b><i>abc</i></b><i>d<br></i></div>", "<div><b><i>abc</i></b><i>d</i><br></div>"]
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]</i><br></b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]</i><br></b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i><br></b></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><b><i>abc</i></b><i>d<br></i><b><br></b></div>" but got "<div><i style=\\"\\"><b>abc</b>d<br></i><br></div>"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i><br></b></div>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["styleWithCSS","false"],["bold",""],["inserttext","d"]] "<div><b><i>abc[]<br></i><br></b></div>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertparagraph",""],["inserttext","c"],["inserttext","d"],["inserttext","e"]] "<div>ab[c<b><i>de]f</i></b>ghi</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>ab</div><div><b><i>cdef</i></b>ghi</div>" but got "<div>ab</div><div>cde<i style=\\"font-weight:bold\\">f</i>ghi</div>"
[FAIL] [["insertlinebreak",""],["inserttext","c"],["inserttext","d"],["inserttext","e"]] "<div>ab[c<b><i>de]f</i></b>ghi</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>ab<br><b><i>cdef</i></b>ghi</div>" but got "<div>ab<br>cde<i style=\\"font-weight:bold\\">f</i>ghi</div>"
[FAIL] [["styleWithCSS","false"],["italic",""],["bold",""],["insertText","b"]] "a[]c" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "a<i><b>b</b></i>c" but got "a<b><i>b</i></b>c"
[FAIL] [["delete",""],["inserttext","g"],["inserttext","h"]] "<div>{abc</div><div>def</div>}" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "gh" not in array ["<div>gh</div>", "<div>gh<br></div>"]
[FAIL] [["delete",""],["inserttext","g"],["inserttext","h"]] "<div><b>{abc</b></div><div>def</div>}" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<b>gh</b>" not in array ["<div><b>gh</b></div>", "<div><b>gh<br></b></div>", "<div><b>gh</b><br></div>"]
[FAIL] [["delete",""],["inserttext","Y"]] "<b>X[]<span contenteditable=false>abc</span></b><i>def</i>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "Y<b><span contenteditable=\\"false\\">abc</span></b><i>def</i>" not in array ["<b>Y<span contenteditable=\\"false\\">abc</span></b><i>def</i>", "<b>Y</b><b><span contenteditable=\\"false\\">abc</span></b><i>def</i>"]
[FAIL] [["forwarddelete",""],["inserttext","Y"]] "<b>[]X<span contenteditable=false>abc</span></b><i>def</i>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "Y<b><span contenteditable=\\"false\\">abc</span></b><i>def</i>" not in array ["<b>Y<span contenteditable=\\"false\\">abc</span></b><i>def</i>", "<b>Y</b><b><span contenteditable=\\"false\\">abc</span></b><i>def</i>"]
Harness: the test ran to completion.
